requirements.txt
##First packages needed
python-dotenv
uvicorn[standard]
requests
streamlit

##For Cohere part we will need that
cohere
fastAPI

##For RAG functionality
sentence-transformers
numpy
scikit-learn